pyproject.toml
[tool.poetry]
name = "CDMO Project - Mariani"
version = "0.1.0"
description = ""
authors = ["Manuel Mariani <manuel.mariani2@studio.unibo.it>"]

[tool.poetry.dependencies]
python = "3.8.5"
numpy = "1.19.2"
matplotlib = "3.3.2"
plotly = "4.14.3"
pandas = "1.1.5"
PyQt5 = "^5.15.4"
rich = "^10.12.0"
Pillow = "8.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88